Remove runtime hack to load themes (#520)
* Use antigen-compat file rather than doing it at runtime
* Update tests, remove references to _CACHE_BUNDLE

+++ src/boot.zsh
@@ -4,13 +4,17 @@
 # Homepage: http://antigen.sharats.me
 # License: MIT License <mitl.sharats.me>
 
-ANTIGEN_CACHE="${ANTIGEN_CACHE:-${ADOTDIR:-$HOME/.antigen}/init.zsh}"
+if [[ $ANTIGEN_CACHE != false ]]; then
+  ANTIGEN_CACHE="${ANTIGEN_CACHE:-${ADOTDIR:-$HOME/.antigen}/init.zsh}"
 
-for config in $ANTIGEN_CHECK_FILES; do
-  if [[ "$config" -nt "$config.zwc" ]]; then
-    { zcompile "$config" } &!
-    [[ -f "$ANTIGEN_CACHE" ]] && rm -f "$ANTIGEN_CACHE"
-  fi
-done
+  for config in $ANTIGEN_CHECK_FILES; do
+    if [[ "$config" -nt "$config.zwc" ]]; then
+      { zcompile "$config" } &!
+      [[ -f "$ANTIGEN_CACHE" ]] && rm -f "$ANTIGEN_CACHE"
+    fi
+  done
 
-[[ -f $ANTIGEN_CACHE && ! $_ANTIGEN_CACHE_LOADED == true ]] && source "$ANTIGEN_CACHE" && return 0;
+  [[ -f $ANTIGEN_CACHE && ! $_ANTIGEN_CACHE_LOADED == true ]] && source "$ANTIGEN_CACHE" && return 0;
+fi
+
+
